The moved function only has a very specific usecase in testing. So doesn't belong in generic production library.
