+++ kong/tools/io.lua
@@ -102,41 +102,6 @@ function _M.file_size(path)
   return size
 end
 
-function _M.retrieve_files(dir, options)
--- TODO : this function is dead code, remove?
-  local fs = require "luarocks.fs"
-  local pattern = options.file_pattern
-  local exclude_dir_patterns = options.exclude_dir_patterns
-
-  if not pattern then pattern = "" end
-  if not exclude_dir_patterns then exclude_dir_patterns = {} end
-  local files = {}
-
-  local function tree(dir)
-    for _, file in ipairs(fs.list_dir(dir)) do
-      local f = path:join(dir, file)
-      if fs.is_dir(f) then
-        local is_ignored = false
-        for _, pattern in ipairs(exclude_dir_patterns) do
-          if string.match(f, pattern) then
-            is_ignored = true
-            break
-          end
-        end
-        if not is_ignored then
-          tree(f)
-        end
-      elseif fs.is_file(f) and string.match(file, pattern) ~= nil then
-        table.insert(files, f)
-      end
-    end
-  end
-
-  tree(dir)
-
-  return files
-end
-
 --- Load a yaml configuration file.
 -- The return config will get 2 extra fields; `pid_file` of the nginx process 
 -- and `dao_config` as a shortcut to the dao configuration
